Extract reporting query contract modules
