+++ src/app/contracts/reporting_query_search.py
@@ -1,0 +1,210 @@
+from datetime import datetime
+from typing import Any
+
+from pydantic import BaseModel, Field
+
+from app.contracts.reporting_query_examples import (
+    REPORT_JOB_LIST_FILTERS_EXAMPLE,
+    REPORT_JOB_LIST_RESPONSE_EXAMPLE,
+)
+
+__all__ = [
+    "ReportJobListFilters",
+    "ReportJobListItem",
+    "ReportJobListResponse",
+]
+
+
+class ReportJobListFilters(BaseModel):
+    tenant_id: str | None = Field(
+        default=None,
+        alias="tenantId",
+        description="Tenant filter used to isolate jobs for one tenant scope.",
+        examples=["tenant-sg"],
+    )
+    region: str | None = Field(
+        default=None,
+        alias="region",
+        description="Region filter used to isolate jobs for one operating region.",
+        examples=["APAC"],
+    )
+    status: str | None = Field(
+        default=None,
+        alias="status",
+        description="Current job-status filter.",
+        examples=["accepted"],
+    )
+    report_type: str | None = Field(
+        default=None,
+        alias="reportType",
+        description="Report-type filter for the job search.",
+        examples=["portfolio_review"],
+    )
+    portfolio_id: str | None = Field(
+        default=None,
+        alias="portfolioId",
+        description="Portfolio identifier contained in the submitted portfolio scope.",
+        examples=["PB_SG_GLOBAL_BAL_001"],
+    )
+    as_of_date: str | None = Field(
+        default=None,
+        alias="asOfDate",
+        description="Business as-of date filter for the report request.",
+        examples=["2026-04-22"],
+    )
+    idempotency_key: str | None = Field(
+        default=None,
+        alias="idempotencyKey",
+        description="Idempotency key filter for duplicate-request diagnostics.",
+        examples=["portfolio-review-PB_SG_GLOBAL_BAL_001-2026-04-22"],
+    )
+    correlation_id: str | None = Field(
+        default=None,
+        alias="correlationId",
+        description="Correlation identifier filter for end-to-end operational tracing.",
+        examples=["corr-portfolio-review-1"],
+    )
+    created_from: datetime | None = Field(
+        default=None,
+        alias="createdFrom",
+        description="Inclusive lower UTC bound for job creation time.",
+        examples=["2026-04-22T00:00:00Z"],
+    )
+    created_to: datetime | None = Field(
+        default=None,
+        alias="createdTo",
+        description="Inclusive upper UTC bound for job creation time.",
+        examples=["2026-04-23T00:00:00Z"],
+    )
+    limit: int = Field(
+        default=25,
+        alias="limit",
+        description="Maximum number of jobs returned by this bounded search.",
+        examples=[25],
+    )
+
+    model_config = {"populate_by_name": True}
+
+
+class ReportJobListItem(BaseModel):
+    report_job_id: str = Field(
+        ...,
+        alias="reportJobId",
+        description="Opaque durable report job identifier.",
+        examples=["rjob_83ca965c50334c40a17d2b8cc94873a5"],
+    )
+    report_request_id: str = Field(
+        ...,
+        alias="reportRequestId",
+        description="Opaque durable report request identifier linked to the job.",
+        examples=["rrq_4f7c85b39f7d4e7b8d0bb420d34a1d2c"],
+    )
+    report_type: str = Field(
+        ...,
+        alias="reportType",
+        description="Report type handled by the job.",
+        examples=["portfolio_review"],
+    )
+    tenant_id: str = Field(
+        ...,
+        alias="tenantId",
+        description="Tenant identifier captured when the request was created.",
+        examples=["tenant-sg"],
+    )
+    region: str = Field(
+        ...,
+        alias="region",
+        description="Operating region captured when the request was created.",
+        examples=["APAC"],
+    )
+    portfolio_scope: dict[str, Any] = Field(
+        ...,
+        alias="portfolioScope",
+        description="Submitted portfolio scope for the report job.",
+        examples=[{"portfolio_ids": ["PB_SG_GLOBAL_BAL_001"]}],
+    )
+    as_of_date: str = Field(
+        ...,
+        alias="asOfDate",
+        description="Business as-of date submitted for the report job.",
+        examples=["2026-04-22"],
+    )
+    status: str = Field(
+        ...,
+        alias="status",
+        description="Current product-safe report job status.",
+        examples=["accepted"],
+    )
+    failure_category: str | None = Field(
+        default=None,
+        alias="failureCategory",
+        description="Machine-readable failure category when the job failed or was cancelled.",
+        examples=[None],
+    )
+    current_step: str = Field(
+        ...,
+        alias="currentStep",
+        description="Current lifecycle step for support diagnostics.",
+        examples=["accepted"],
+    )
+    retry_eligible: bool = Field(
+        ...,
+        alias="retryEligible",
+        description="Whether retry or replay is currently permitted.",
+        examples=[False],
+    )
+    cancel_requested: bool = Field(
+        ...,
+        alias="cancelRequested",
+        description="Whether cancellation has been requested and recorded.",
+        examples=[False],
+    )
+    idempotency_key: str = Field(
+        ...,
+        alias="idempotencyKey",
+        description="Caller-supplied idempotency key associated with the job.",
+        examples=["portfolio-review-PB_SG_GLOBAL_BAL_001-2026-04-22"],
+    )
+    correlation_id: str = Field(
+        ...,
+        alias="correlationId",
+        description="Correlation identifier captured when the request was created.",
+        examples=["corr-portfolio-review-1"],
+    )
+    created_at: datetime = Field(
+        ...,
+        alias="createdAt",
+        description="UTC timestamp when the job was created.",
+        examples=["2026-04-22T09:00:00Z"],
+    )
+    updated_at: datetime = Field(
+        ...,
+        alias="updatedAt",
+        description="UTC timestamp when the job was last updated.",
+        examples=["2026-04-22T09:00:00Z"],
+    )
+
+    model_config = {"populate_by_name": True}
+
+
+class ReportJobListResponse(BaseModel):
+    count: int = Field(
+        ...,
+        alias="count",
+        description="Number of jobs returned in this bounded response.",
+        examples=[1],
+    )
+    applied_filters: ReportJobListFilters = Field(
+        ...,
+        alias="appliedFilters",
+        description="Normalized filters applied to the job search.",
+        examples=[REPORT_JOB_LIST_FILTERS_EXAMPLE],
+    )
+    items: list[ReportJobListItem] = Field(
+        ...,
+        alias="items",
+        description="Bounded list of support-safe report job summaries.",
+        examples=[REPORT_JOB_LIST_RESPONSE_EXAMPLE["items"]],
+    )
+
+    model_config = {"populate_by_name": True}
