+++ src/app/contracts/reporting_query_status.py
@@ -1,0 +1,89 @@
+from datetime import datetime
+
+from pydantic import BaseModel, Field
+
+__all__ = [
+    "ReportJobStatusEventsResponse",
+    "ReportStatusEvent",
+]
+
+
+class ReportStatusEvent(BaseModel):
+    status_event_id: str = Field(
+        ...,
+        description="Opaque append-only status event identifier.",
+        examples=["rse_d7e9c3b87d864b098997d4fe5bd2de2a"],
+    )
+    report_job_id: str = Field(
+        ...,
+        description="Report job identifier associated with this event.",
+        examples=["rjob_83ca965c50334c40a17d2b8cc94873a5"],
+    )
+    from_status: str | None = Field(
+        default=None,
+        description="Previous status when this event is a transition.",
+        examples=[None],
+    )
+    to_status: str = Field(
+        ...,
+        description="New status recorded by this event.",
+        examples=["accepted"],
+    )
+    event_type: str = Field(
+        ...,
+        description="Machine-readable lifecycle event type.",
+        examples=["job_accepted"],
+    )
+    message: str | None = Field(
+        default=None,
+        description="Support-safe lifecycle event message.",
+        examples=["Portfolio review report job accepted."],
+    )
+    actor: str = Field(
+        ...,
+        description="Actor or system principal associated with this event.",
+        examples=["advisor-123"],
+    )
+    created_at: datetime = Field(
+        ...,
+        description="UTC timestamp when this event was appended.",
+        examples=["2026-04-22T09:00:00Z"],
+    )
+    correlation_id: str = Field(
+        ...,
+        description="Correlation identifier associated with this event.",
+        examples=["corr-portfolio-review-1"],
+    )
+    trace_id: str = Field(
+        ...,
+        description="Distributed trace identifier associated with this event.",
+        examples=["4bf92f3577b34da6a3ce929d0e0e4736"],
+    )
+
+
+class ReportJobStatusEventsResponse(BaseModel):
+    report_job_id: str = Field(
+        ...,
+        description="Report job identifier whose event history is returned.",
+        examples=["rjob_83ca965c50334c40a17d2b8cc94873a5"],
+    )
+    events: list[ReportStatusEvent] = Field(
+        ...,
+        description="Append-only lifecycle events ordered by creation time.",
+        examples=[
+            [
+                {
+                    "statusEventId": "rse_d7e9c3b87d864b098997d4fe5bd2de2a",
+                    "reportJobId": "rjob_83ca965c50334c40a17d2b8cc94873a5",
+                    "fromStatus": None,
+                    "toStatus": "accepted",
+                    "eventType": "job_accepted",
+                    "message": "Portfolio review report job accepted.",
+                    "actor": "advisor-123",
+                    "createdAt": "2026-04-22T09:00:00Z",
+                    "correlationId": "corr-portfolio-review-1",
+                    "traceId": "4bf92f3577b34da6a3ce929d0e0e4736",
+                }
+            ]
+        ],
+    )
