+++ wiki/Architecture.md
@@ -10,10 +10,10 @@
 - consolidated architecture and quality-baseline docs under
   [docs/architecture.md](../docs/architecture.md) and
   [quality/architecture_rules.md](../quality/architecture_rules.md)
-- current enterprise-hardening evidence records `src/app/contracts/reporting_query.py`
-  as the largest residual source file at 664 script-counted lines after performance horizon row
-  assembly moved into `src/app/services/performance_workspace_horizon_rows.py`; the blocking
-  source-file threshold is ratcheted to 664 script-counted lines, and the repository longest-function baseline
+- current enterprise-hardening evidence records `src/app/contracts/reporting_batches.py`
+  as the largest residual source file at 662 script-counted lines after reporting query status,
+  search, lineage, and example contracts moved behind the existing reporting query facade; the blocking
+  source-file threshold is ratcheted to 662 script-counted lines, and the repository longest-function baseline
   remains held at the enforced 49-line AST span
 
 ## Route-family map
